llm/rag/requirements.txt
faiss-cpu
gradio==5.25.1
huggingface_hub==0.30.2
ipython==8.12.3
langchain==0.3.23
langchain_core==0.3.51
langchain_google_genai==2.1.2
pandas==2.2.3
protobuf==6.30.2
python-dotenv==1.1.0
pypdf